backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic